Refactor slice utils package structure
