+++ internal/catalogmetadata/filter/bundle_predicates.go
@@ -6,10 +6,10 @@ import (
 	"github.com/operator-framework/operator-registry/alpha/declcfg"
 
 	"github.com/operator-framework/operator-controller/internal/bundleutil"
-	slicesutil "github.com/operator-framework/operator-controller/internal/util/slices"
+	filterutil "github.com/operator-framework/operator-controller/internal/util/filter"
 )
 
-func InMastermindsSemverRange(semverRange *mmsemver.Constraints) slicesutil.Predicate[declcfg.Bundle] {
+func InMastermindsSemverRange(semverRange *mmsemver.Constraints) filterutil.Predicate[declcfg.Bundle] {
 	return func(b declcfg.Bundle) bool {
 		bVersion, err := bundleutil.GetVersion(b)
 		if err != nil {
@@ -27,7 +27,7 @@ func InMastermindsSemverRange(semverRange *mmsemver.Constraints) slicesutil.Pred
 	}
 }
 
-func InAnyChannel(channels ...declcfg.Channel) slicesutil.Predicate[declcfg.Bundle] {
+func InAnyChannel(channels ...declcfg.Channel) filterutil.Predicate[declcfg.Bundle] {
 	return func(bundle declcfg.Bundle) bool {
 		for _, ch := range channels {
 			for _, entry := range ch.Entries {
